refactor: split storage and sorting modules
