+++ index.html
@@ -236,7 +236,7 @@
     </div>
 
     <!-- Load JavaScript -->
-    <script src="js/script.js"></script>
+    <script type="module" src="js/script.js"></script>
     
     <!-- Service Worker Registration for PWA -->
     <script>
